Add details to impersonation guards
In order to make the impersonation guards (´canImpersonate´ and ´canBeImpersonated`)
as useful as possible, the opposite user (impersonated user for `canImpersonate` and
impersonator for `canBeImpersonated`) is provided as an argument.
The ImpersonateManager checks if these guards are in place and checks if they return
true. If not, the newly added `CannotBeImpersonatedException` and
`CannotImpersonateException` will be thrown.
A `forceTake` can be used to omit these checks and definitely impersonate a user
regardless the guards.